Move exportOptions.plist generator functions to export_options.go file.
